requirements.txt
Flask==3.0.3
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
Flask-Cors
pymongo
Flask-RESTful
pydantic
Flask-Mail

annotated-types==0.6.0
google-api-core==2.18.0
googleapis-common-protos==1.63.0
urllib3==2.2.1
python-dotenv
# cashfree_pg
boto3
razorpay